requirements.txt
plotly==4.14.3
numpy==1.19.5
pandas==1.2.2
streamlit==0.80.0
matplotlib==3.3.4
Pillow==8.2.0
scikit_learn==0.24.1
